Adding paymentIntentId to the backend

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,12 @@
-import { AfterViewInit, Component, ElementRef, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
+import {
+  AfterViewInit,
+  Component,
+  ElementRef,
+  Input,
+  OnDestroy,
+  OnInit,
+  ViewChild,
+} from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { NavigationExtras, Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
@@ -12,7 +20,7 @@ declare var Stripe;
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
-  styleUrls: ['./checkout-payment.component.scss']
+  styleUrls: ['./checkout-payment.component.scss'],
 })
 export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   // 19.248
@@ -20,9 +28,9 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
 
   //#region 21.266
   //viewchild to get access to html template reference
-  @ViewChild('cardNumber', {static: true}) cardNumberElement: ElementRef;
-  @ViewChild('cardExpiry', {static: true}) cardExpiryElement: ElementRef;
-  @ViewChild('cardCvc', {static: true}) cardCvcElement: ElementRef;
+  @ViewChild('cardNumber', { static: true }) cardNumberElement: ElementRef;
+  @ViewChild('cardExpiry', { static: true }) cardExpiryElement: ElementRef;
+  @ViewChild('cardCvc', { static: true }) cardCvcElement: ElementRef;
   stripe: any;
   cardNumber: any;
   cardExpiry: any;
@@ -36,15 +44,18 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
     private basketService: BasketService,
     private checkoutService: CheckoutService,
     private toastr: ToastrService,
-    private router: Router) { }
+    private router: Router
+  ) {}
 
-  //#region 21.266 Adding stripe elements into component 
+  //#region 21.266 Adding stripe elements into component
   //instead of ngoninit it needs to be ngAfterViewInit
   /*ngOnInit(): void {
   }*/
 
   ngAfterViewInit() {
-    this.stripe = Stripe('pk_test_51J9chTGfXLXdWZmRfaiIwJuWmLyhWur1Py8V2ArZ48h5vWW8UviQLUf7rJMuqDa6uaXhr5uXqwsj8kKVgkKtfyyL00OoldXCsv');
+    this.stripe = Stripe(
+      'pk_test_51J9chTGfXLXdWZmRfaiIwJuWmLyhWur1Py8V2ArZ48h5vWW8UviQLUf7rJMuqDa6uaXhr5uXqwsj8kKVgkKtfyyL00OoldXCsv'
+    );
     const elements = this.stripe.elements();
 
     this.cardNumber = elements.create('cardNumber');
@@ -68,11 +79,11 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
     this.cardExpiry.destroy();
     this.cardCvc.destroy();
   }
-  //#endregion 
+  //#endregion
 
   //#region 21.267 Displaying validation errors beneith the fields -> checkout-payment.html
-  onChange({error}) {
-    if(error) {
+  onChange({ error }) {
+    if (error) {
       this.cardErrors = error.message;
     } else {
       this.cardErrors = null;
@@ -84,25 +95,47 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   submitOrder() {
     const basket = this.basketService.getCurrentBasketValue();
     const orderToCreate = this.getOrderToCreate(basket);
-    this.checkoutService.creatOrder(orderToCreate).subscribe((order: IOrder) => {
-      this.toastr.success('Order created successfully');
-      this.basketService.deleteLocalBasket(basket.id);
-      const navigationExtras: NavigationExtras = {state: order};
-      this.router.navigate(['checkout/success'], navigationExtras);
-    }, error => {
-      this.toastr.error(error.message);
-      console.log(error);
-    })
+    this.checkoutService.creatOrder(orderToCreate).subscribe(
+      (order: IOrder) => {
+        this.toastr.success('Order created successfully');
+        //#region 21.269 Submiting the payment -> Order.cs
+        // adding stripe funcionality
+        // confirm card payment
+        this.stripe.confirmCardPayment(basket.clientSecret, {
+          payment_method: {
+            card: this.cardNumber,
+            billing_details: {
+              name: this.checkoutForm.get('paymentForm').get('nameOnCard').value
+            }
+          }
+        }).then(result => {
+          console.log(result);
+            // in case that payment is successuful
+            if (result.paymentIntent) {
+              this.basketService.deleteLocalBasket(basket.id);
+              const navigationExtras: NavigationExtras = {state: order};
+              this.router.navigate(['checkout/success'], navigationExtras);
+            } else {
+              this.toastr.error(result.error.message);
+            }
+          });
+        //#endregion
+      },
+      (error) => {
+        this.toastr.error(error.message);
+        console.log(error);
+      }
+    );
   }
-
 
   private getOrderToCreate(basket: IBasket) {
     return {
       basketId: basket.id,
-      deliveryMethodId: +this.checkoutForm.get('deliveryForm').get('deliveryMethod').value,
-      shipToAddress: this.checkoutForm.get('addressForm').value
+      deliveryMethodId: +this.checkoutForm
+        .get('deliveryForm')
+        .get('deliveryMethod').value,
+      shipToAddress: this.checkoutForm.get('addressForm').value,
     };
   }
   //#endregion
-
 }
